correcao wallet valores

+++ src/main/java/com/finances/Finances/V1/dto/wallet/WalletRequest.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -14,8 +15,8 @@ import javax.validation.constraints.NotNull;
 public class WalletRequest {
 
     @NotNull
-    private Double currentBalance;
+    private BigDecimal currentBalance;
 
     @NotNull
-    private Double monthlyExpense;
+    private BigDecimal monthlyExpense;
 }
