+++ src/main/java/com/finances/Finances/V1/model/Wallet.java
@@ -3,10 +3,8 @@ package com.finances.Finances.V1.model;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -23,7 +21,7 @@ public class Wallet {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private UUID id;
 
-    private double currentBalance;
+    private BigDecimal currentBalance;
 
-    private double monthlyExpense;
+    private BigDecimal monthlyExpense;
 }
